Added the RequestUpdateUniversityEvent and RequestUpdateUniversityEvent and implementation for updateUniversity

+++ pms-webserver/src/main/java/de/hfu/pms/exceptions/UniversityNotFoundException.java
@@ -2,7 +2,7 @@ package de.hfu.pms.exceptions;
 
 public class UniversityNotFoundException extends RuntimeException {
 
-    public UniversityNotFoundException(Integer id) {
+    public UniversityNotFoundException(Long id) {
         super("Could not find university with id " + id);
     }
 }
